// src/bootgrid/interfaces/bootgrid-item.ts
export interface BootgridItem {
    id: string; // unique id for grid item, auto generated by default
    startPosition: number; // start position in grid
    size: number; // item size
    smSize?: number; // item size
    mdSize?: number; // item size
    lgSize?: number; // item size
    xlSize?: number; // item size
    position: number; // item position in grid
    smPosition?: number; // item position in grid
    mdPosition?: number; // item position in grid
    lgPosition?: number; // item position in grid
    xlPosition?: number; // item position in grid
    componentSelector?: string; // optional component to render in grid item
    focus?: boolean; // focus on item
    autoHeight: boolean; // should height be set to auto
}